code/requirements.txt
azure-functions==1.18.0
streamlit==1.30.0
scipy==1.12.0
transformers==4.38.1
python-dotenv==1.0.1
azure-ai-formrecognizer==3.3.2
azure-storage-blob==12.19.0
azure-identity==1.15.0
Flask==3.0.2
openai==1.6.1
azure-storage-blob==12.19.0
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.24
azure-ai-formrecognizer==3.3.2
requests==2.31.0
tiktoken==0.6.0
azure-storage-queue==12.9.0
beautifulsoup4==4.12.3
fake-useragent==1.4.0
chardet==5.2.0
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-search-documents==11.4.0
opencensus-ext-azure==1.1.13
azure-ai-contentsafety==1.0.0
python-docx==1.1.0
azure-keyvault-secrets==4.4.*
openpyxl
llama-index
duckduckgo-search
psycopg2-binary